Renaming gcloud._localstack as gcloud._helpers.
This is in preparation for adding more core helper classes.
